% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAciPlots.R
\name{aciPlot2}
\alias{aciPlot2}
\title{aciPlot2}
\usage{
aciPlot2(
  D,
  cutname = "Evenson",
  cats = c("Sedentary", "Light", "Moderate", "Vigorous"),
  plot = TRUE,
  select = Wkdy \%in\% 0:6,
  ...
)
}
\arguments{
\item{D}{}

\item{cutname}{}

\item{cats}{}

\item{plot}{}

\item{select}{}

\item{...}{}
}
\description{
aciPlot2
}
