% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_ggir.R
\name{afterGGIR}
\alias{afterGGIR}
\title{afterGGIR}
\usage{
afterGGIR(datadir, sib = FALSE, verbose = TRUE)
}
\arguments{
\item{datadir}{Name of the raw data folder}

\item{sib}{Set TRUE to include a table of sustained inactivity periods from GGIR output}

\item{verbose}{set FALSE to suppress bloody messages}
}
\value{

}
\description{
afterGGIR
}
