% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAciPlots.R
\name{ctaplot}
\alias{ctaplot}
\title{Ctaplot}
\usage{
ctaplot(x, FUN = mean, XLAB = "Kellaaeg", YLAB = "Aktiivsus", MAIN = NULL)
}
\arguments{
\item{x}{}

\item{FUN}{}

\item{XLAB}{}

\item{YLAB}{}

\item{MAIN}{}
}
\description{
Ctaplot
}
