% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_ggir.R
\name{decodeGGIR}
\alias{decodeGGIR}
\title{decode GGIR}
\usage{
decodeGGIR(x, folder = NULL, minimize = FALSE)
}
\arguments{
\item{x}{}

\item{folder}{}

\item{minimize}{}
}
\value{

}
\description{
decode GGIR
}
