% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fb.R
\name{do_fb}
\alias{do_fb}
\title{Simple feedback}
\usage{
do_fb(DIR = "rda", ij = TRUE, UNLINK = TRUE, UNLINK.ALL = TRUE)
}
\arguments{
\item{DIR}{}

\item{ij}{}

\item{UNLINK}{}

\item{UNLINK.ALL}{}
}
\value{
nothing
}
\description{
Simple feedback
}
