% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGGIR1.R
\name{doGGIR1}
\alias{doGGIR1}
\title{doGGIR one file at a time}
\usage{
doGGIR1(indir, III = NULL, SINK = TRUE)
}
\arguments{
\item{indir}{Folder name (*.bin files)}

\item{III}{Integer vector - rank number[s] of files to be processed. Can be NULL (default) - in that case, all files in the folder will included}
}
\value{
system time
}
\description{
doGGIR one file at a time
}
