% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsamp.R
\name{find_windows}
\alias{find_windows}
\title{Find_windows}
\usage{
find_windows(
  E,
  A = "rda",
  mins = 30,
  E.code = "id",
  E.time = "time",
  extramins = 5,
  cuts = c(sed = 0, lig = 52, mod = 192, vig = 540)
)
}
\arguments{
\item{E}{data frame with events data}

\item{A}{activity data folder (the result of afterGGIR, defaults to "rda")}

\item{mins}{width of the activity window (in minutes) i.e. how much time before the event to consider}

\item{extramins}{extra minutes between the end of activity window and the event (a negative number might be used if activity AFTER rather than BEFORE the event is of interest)}

\item{cuts}{cutoffs for activity categories: a named vector of lower thresholds}

\item{codevar}{name of the ID variable in the data frame E (unique identifier of each person)}

\item{timevar}{name of the `time` variable in the data frame E (must be of POSIXt type or converted before analysis)}
}
\value{
A data frame (E) with extra variables added
}
\description{
Given a folder of preprocessed (with afterGGIR) files and a data table with events,
find average and SD of activity (g) preceding each event and time in intensity categories.
For example, E may include a number of experience sampling data points, with variable called 
"time" indicating the time of the "beep". With default values, the activity from 35 to 5 minutes
before the "beep" is analyzed, and mean activity, as well as time in intensity zones (defined by "cuts")
is reported.
}
