% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ggir.R
\name{Load}
\alias{Load}
\title{Load}
\usage{
Load(
  file,
  to.list = TRUE,
  Unlist = TRUE,
  spray = FALSE,
  delete.parent = FALSE,
  Attach = FALSE
)
}
\arguments{
\item{file}{}

\item{to.list}{}

\item{Unlist}{}

\item{spray}{}

\item{delete.parent}{}

\item{Attach}{}
}
\value{
value
}
\description{
Load
}
