% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ggir.R
\name{loadg}
\alias{loadg}
\title{load ggir output from 'results' folder}
\usage{
loadg(FN, OD = "", browse = FALSE)
}
\arguments{
\item{FN}{file name or number}

\item{OD}{original directory name}

\item{browse}{FALSE}
}
\value{

}
\description{
Lorem ipsum
}
