% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reintegrate.R
\name{reintegrate2}
\alias{reintegrate2}
\title{"Reintegrate i.e.}
\usage{
reintegrate2(x, TO = 60, ...)
}
\arguments{
\item{x}{an afterGGIR object}

\item{TO}{the desired resolution ("epoch") in seconds}

\item{...}{currently ignored}
}
\description{
"Reintegrate i.e.
}
