% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resave.R
\name{resave}
\alias{resave}
\title{Resave}
\usage{
resave(IN, OUT)
}
\arguments{
\item{IN}{}

\item{OUT}{}
}
\value{

}
\description{
Re-saves a text file (which can be a *.bin file from GeneActiv) line by line.
This will (in many cases!) get rid of code table conflicts
}
