% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAciPlots.R
\name{toacc}
\alias{toacc}
\title{Convert an afterGGIR object to accelerate::acc}
\usage{
toacc(x)
}
\arguments{
\item{x}{}
}
\value{
an accelerate::acc object
}
\description{
Convert an afterGGIR object to accelerate::acc
}
