% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_ggir.R
\name{decodeGGIR}
\alias{decodeGGIR}
\title{decode GGIR}
\usage{
decodeGGIR(x, folder = NULL, minimize = FALSE, sib = FALSE)
}
\arguments{
\item{x}{Number or name of the file to be read in}

\item{folder}{Name of the raw data folder}

\item{minimize}{set TRUE to forget some of the data (mean light, temperature etc)}

\item{sib}{set TRUE to report sustained inactivity periods (attr(res, "sib"))}
}
\description{
Lorem ipsum
}
